+++ libs/astarte_rpc/lib/astarte_rpc/config/astarte_services.ex
@@ -25,7 +25,7 @@ defmodule Astarte.RPC.Config.AstarteServices do
 
   use Skogsra.Type
 
-  @all_services [:astarte_data_updater_plant, :astarte_vmq_plugin]
+  @all_services [:astarte_data_updater_plant, :astarte_realm_management, :astarte_vmq_plugin]
   @all_services_mapset MapSet.new(@all_services)
 
   @impl Skogsra.Type