+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- New environment variables for trigger notifications between realm management repliacs and realm management -> pairing. These variables are currently being used only by realm management
+  - `REALM_MANAGEMENT_CLUSTERING_KUBERNETES_SELECTOR`. The Endpoint label to query to get other realm management instances. Defaults to `app=astarte-realm-management`.
+
 ## [1.3.0-rc.0] - 2025-11-21
 
 ### Added